Minimize the time lock is held in case of WS message broadcast
In the broadcast case, the mutex is locked and then all clients are
looped through, performing a network write (client.WriteMessage) for
each one.

If one client is slow or has a poor connection, WriteMessage could block
for a significant amount of time. Because the mutex is held during this
entire loop, no new clients can be registered or unregistered until the
broadcast to all clients is complete. This can lead to performance
issues and contention.

A better approach is to minimize the time the lock is held. This can be
achieve by making a copy of the client list under the lock, and then
iterating over the copy to send the messages.